Fixed issue when MILP problem is unfeasible
